requirements.txt
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.10.0
Django==3.2.6
django-appconf==1.0.5
django-bootstrap-v5==1.0.7
django-imagekit==4.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
pilkit==2.0
pycodestyle==2.7.0
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
soupsieve==2.3
sqlparse==0.4.1
toml==0.10.2
